+++ apps/web/src/features/project/model/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './template';
