+++ apps/web/src/features/project/model/constants/template.ts
