+++ apps/web/src/features/project/ui/index.ts
@@ -1,6 +1,7 @@
 export { default as ChangeProjectModal } from './ChangeProjectModal';
-export { default as CreateProjectModal } from './CreateProjectModal';
 export { default as ExpulsionProjectModal } from './ExpulsionProjectModal';
 export { default as InviteProjectModal } from './InviteProjectModal';
 export { default as LeaveProjectModal } from './LeaveProjectModal';
 export { default as ProjectManageModal } from './ProjectManageModal';
+export { default as CreateGroupBoardModal } from './CreateGroupBoardModal';
+export { default as ProjectMemberManageModal } from './ProjectMemberManageModal';
